RelFinder is now working with the current Version of Flex 3.5
The Input fields now have a delay of 0.75 seconds before they start the search for suggestions with the actual input. This reduces the number of search queries and it's faster to wait 0.75 seconds after the input was completed, then to wait for all queries requested by every new char input.